Added a PM7 library generated by thermoEstimator
I am not sure if this is the right  place to store QM thermo libraries generated by RMG or thermoEstimator, but it should be stored somewhere to avoid duplicate efforts.

This library has 6597 species obtained from a dictionary file of a JP10 job. QM settings are:
"""
quantumMechanics(
    software='mopacPM7',
    fileStore='/scratch/keceli/QMfiles',
    scratchDirectory = None, # not currently used
    onlyCyclics = True,
    maxRadicalNumber = 0,
)
"""
It took 157 minutes with 48 processsors.